video: msm: mdss: Restructure panel driver for SONY devices
This driver supports Sony devices, starting from the old Xperia Z1
and E3 to the newest Xperia Z5.

At the beginning, it was relatively big, but still okay to have it
in basically just one *.c file, but now it has grown and it has
become really huge and, as a result of this, understanding what
was going on in the driver had become very difficult.

This patch is doing some restructuration to the driver along
with some cleanups and additions, like reworked panel detection
and cleaned up extra GPIOs handling (which are usually used only
in mid-end and low-end devices, currently Yukon and Kanuti boards)